+++ src/Core/src/Exception/UnparsableResponse.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace AsyncAws\Core\Exception;
 
-class ParseResponse extends \RuntimeException implements Exception
+class UnparsableResponse extends \RuntimeException implements Exception
 {
 }